SET ACCES for StreamsController@edit@update@destroy@create@store #2

+++ resources/views/streams/index.blade.php
@@ -15,10 +15,11 @@
     </aside>
 
     <div class="container stream">
-
-        <p class="text-right">
-            <a href="{{ action('StreamsController@create') }}" class="btn btn-primary"><i class="fa fa-plus"></i> Ajouter un stream </a>
-        </p>
+        @if (Auth::user()->admin == 1)
+            <p class="text-right">
+                <a href="{{ action('StreamsController@create') }}" class="btn btn-primary"><i class="fa fa-plus"></i> Ajouter un stream </a>
+            </p>
+        @endif
 
         <table width="100%" color="black" class="table table-striped">
             <tbody>
@@ -28,21 +29,24 @@
                         <span class="stream__title">
                             <a href="{{ url("/stream/$stream->id ")}}">{{ $stream->name }}</a>
                         </span>
-                        <span class="stream__action">
-                            <a href="{{ action('StreamsController@edit', $stream) }}" class="btn btn-primary"><i class="fa fa-pencil"></i> Editer</a>
-                            <a href="{{ action('StreamsController@destroy', $stream) }}" data-method="delete" data-confirm="Voulez vous vraiment suprimer cette enregistrement ?" class="btn btn-danger"><i class="fa fa-trash"></i> Suprimmer</a>
-                        </span>
+                        @if (Auth::user()->admin == 1)
+                            <span class="stream__action">
+                                <a href="{{ action('StreamsController@edit', $stream) }}" class="btn btn-primary"><i class="fa fa-pencil"></i> Editer</a>
+                                <a href="{{ action('StreamsController@destroy', $stream) }}" data-method="delete" data-confirm="Voulez vous vraiment suprimer cette enregistrement ?" class="btn btn-danger"><i class="fa fa-trash"></i> Suprimmer</a>
+                            </span>
+                        @endif
                         <hr class="stream__hr"/>
 
                         <table>
                             <tbody>
-                                <tr>
-                                    <td valign="top">
-                                        <iframe src="http://player.twitch.tv/?channel={{ $stream->name }}" frameborder="0" scrolling="no" height="242" width="360"></iframe>
-                                    </td>
-                                    <td valign="top">
-                                        <p class="stream__content">{!!  $stream->content !!}</p>
-                                    </td>
+                            <tr>
+                                <td valign="top">
+                                    <iframe src="http://player.twitch.tv/?channel={{ $stream->name }}" frameborder="0" scrolling="no" height="242" width="360"></iframe>
+                                </td>
+                                <td valign="top">
+                                    <p class="stream__content">{!!  $stream->content !!}</p>
+
+                                </td>
                                 </tr>
                             </tbody>
                         </table>
